Remove old code and start replacing it by new code
