+++ inc/LocationBlock.class.hpp
@@ -17,7 +17,6 @@
 # include <string>
 # include <vector>
 # include <map>
-# include "LimitExcept.class.hpp"
 # include "Exception.class.hpp"
 # include "Utility.hpp"
 # include "LimitExceptBlock.class.hpp"
